a little cleanup
