+++ Source/Core/DolphinLib.vcxproj
@@ -70,6 +70,7 @@
     <Text Include="$(BuildInfoTemplate)" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="Subtitles\Helpers.h" />
     <ClInclude Include="Subtitles\WebColors.h" />
     <ClInclude Include="Subtitles\TranslationEntry.h" />
     <ClInclude Include="Subtitles\Subtitles.h" />
