+++ Source/Core/Subtitles/Subtitles.cpp
@@ -1,13 +1,15 @@
 #include "Subtitles.h"
 #include "TranslationEntry.h"
 #include "WebColors.h"
-#include "picojson.h"
+#include "Helpers.h"
+
+#include <picojson.h>
 
 #include <VideoCommon/OnScreenDisplay.h>
-#include "Core/ConfigManager.h"
+#include <Core/ConfigManager.h>
 
-#include <Common\FileUtil.h>
-#include <DiscIO\Filesystem.h>
+#include <Common/FileUtil.h>
+#include <DiscIO/Filesystem.h>
 
 #include <fstream>
 #include <iostream>
@@ -20,17 +22,6 @@ namespace Subtitles
 bool _messageStacksInitialized = false;
 bool _subtitlesInitialized = false;
 std::map<std::string, TranslationEntryGroup> Translations;
-
-void Info(std::string msg)
-{
-  OSD::AddMessage(msg, 2000, OSD::Color::GREEN);
-  INFO_LOG_FMT(SUBTITLES, "{}", msg);
-}
-void Error(std::string err)
-{
-  OSD::AddMessage(err, 2000, OSD::Color::RED);
-  ERROR_LOG_FMT(SUBTITLES, "{}", err);
-}
 
 void DeserializeSubtitlesJson(std::string json)
 {
@@ -72,50 +63,7 @@ void DeserializeSubtitlesJson(std::string json)
       continue;
     }
 
-    u32 color = OSD::Color::CYAN;
-
-    if (Color.is<double>())
-      color = Color.get<double>();
-    else
-    {
-      auto str = Color.to_str();
-      Common::ToLower(&str);
-
-      if (str.starts_with("0x"))
-        //hex string
-        color = std::stoul(str, nullptr, 16);
-      else if (WebColors.count(str) == 1)
-        //html color name
-        color = WebColors[str];
-      else
-        //color noted with 3 or 4 base10 numers (rgb/argb)
-        try //string parsing suucks
-        {
-          // try parse (a)rgb space delimited color
-          u32 a, r, g, b;
-          auto parts = SplitString(str, ' ');
-          if (parts.size() == 4)
-          {
-            a = std::stoul(parts[0], nullptr, 10);
-            r = std::stoul(parts[1], nullptr, 10);
-            g = std::stoul(parts[2], nullptr, 10);
-            b = std::stoul(parts[3], nullptr, 10);
-            color = a << 24 | r << 16 | g << 8 | b;
-          }
-          else if (parts.size() == 3)
-          {
-            a = 255;
-            r = std::stoul(parts[0], nullptr, 10);
-            g = std::stoul(parts[1], nullptr, 10);
-            b = std::stoul(parts[2], nullptr, 10);
-            color = a << 24 | r << 16 | g << 8 | b;
-          }
-        }
-        catch (std::exception x)
-        {
-          Error("Invalid color: " + str);
-        }
-    }
+    u32 color = TryParsecolor(Color, OSD::Color::CYAN);
 
     auto tl = TranslationEntry(FileName.to_str(), Translation.to_str(),
                                Miliseconds.is<double>() ? Miliseconds.get<double>() :
